Fixed another compiler warning
git-svn-id: http://svn.ta3d.org/yuni/trunk@832 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/core/hash/checksum/md5.cpp
@@ -430,7 +430,7 @@ namespace Checksum
 			while (stream.good())
 			{
 				stream.read((char*)buffer, 1024); // note that return value of read is unusable.
-				len = stream.gcount();
+				len = (int) stream.gcount();
 				md5ImplAppend(&state, (const MD5TypeByte*)buffer, len);
 			}
 			md5ImplFinish(&state, digest);
